Cascade upstream invalidation from primitives (Phase 0)
Mirror the cancel-first defense-in-depth pattern: each invalidating
orchestrator primitive (retryTask, recreateTask, retryWorkflow,
recreateWorkflow, forkWorkflow; recreateWorkflowFromFreshBase via
delegation) now calls cascadeInvalidationToDownstream after its local
reset. Production callers reach the primitives directly via
CommandService and workflow-actions.ts, so this is required for the
cascade to fire in production until Phase 2 routes through
applyInvalidation. The cascade always refreshes from the DB before
checking for downstream — memory must not be a negative cache for
a correctness invariant.
