+++ packages/workflow-core/src/__tests__/cross-workflow-cascade.test.ts
@@ -1,136 +1,16 @@
 import { describe, it, expect, beforeEach } from 'vitest';
-import {
-  Orchestrator,
-  type OrchestratorPersistence,
-  type OrchestratorMessageBus,
-} from '../orchestrator.js';
+import { Orchestrator } from '../orchestrator.js';
 import {
   applyInvalidation,
   type InvalidationDeps,
   type InvalidationScope,
 } from '../invalidation-policy.js';
-import type { TaskState, TaskStateChanges, Attempt } from '../task-types.js';
-import type { WorkResponse } from '@invoker/contracts';
-
-// ── In-memory fixtures (mirrors cancel-first-invariant.test.ts) ──
-
-class InMemoryPersistence implements OrchestratorPersistence {
-  workflows = new Map<string, {
-    id: string;
-    name: string;
-    status: string;
-    createdAt: string;
-    updatedAt: string;
-    repoUrl?: string;
-    baseBranch?: string;
-    featureBranch?: string;
-    mergeMode?: 'manual' | 'automatic' | 'external_review';
-  }>();
-  tasks = new Map<string, { workflowId: string; task: TaskState }>();
-  private attempts = new Map<string, Attempt[]>();
-  events: Array<{ taskId: string; eventType: string; payload?: unknown }> = [];
-
-  saveWorkflow(workflow: {
-    id: string;
-    name: string;
-    status: string;
-    repoUrl?: string;
-    baseBranch?: string;
-    featureBranch?: string;
-    mergeMode?: 'manual' | 'automatic' | 'external_review';
-  }): void {
-    const now = new Date().toISOString();
-    this.workflows.set(workflow.id, {
-      ...workflow,
-      repoUrl: workflow.repoUrl ?? 'memory://test-repo',
-      createdAt: (workflow as { createdAt?: string }).createdAt ?? now,
-      updatedAt: (workflow as { updatedAt?: string }).updatedAt ?? now,
-    });
-  }
-  updateWorkflow(): void {}
-  loadWorkflow(workflowId: string): { repoUrl?: string; baseBranch?: string; featureBranch?: string } | undefined {
-    const wf = this.workflows.get(workflowId);
-    return wf
-      ? { repoUrl: wf.repoUrl, baseBranch: wf.baseBranch, featureBranch: wf.featureBranch }
-      : undefined;
-  }
-  saveTask(workflowId: string, task: TaskState): void {
-    this.tasks.set(task.id, { workflowId, task });
-  }
-  getTaskEntry(taskId: string): { workflowId: string; task: TaskState } | undefined {
-    return this.tasks.get(taskId);
-  }
-  updateTask(taskId: string, changes: TaskStateChanges): void {
-    const entry = this.tasks.get(taskId);
-    if (!entry) return;
-    entry.task = {
-      ...entry.task,
-      ...(changes.status !== undefined ? { status: changes.status } : {}),
-      ...(changes.dependencies !== undefined ? { dependencies: changes.dependencies } : {}),
-      config: { ...entry.task.config, ...changes.config },
-      execution: { ...entry.task.execution, ...changes.execution },
-    } as TaskState;
-  }
-  listWorkflows() { return Array.from(this.workflows.values()); }
-  loadTasks(workflowId: string): TaskState[] {
-    return Array.from(this.tasks.values())
-      .filter((e) => e.workflowId === workflowId)
-      .map((e) => e.task);
-  }
-  logEvent(taskId: string, eventType: string, payload?: unknown): void {
-    this.events.push({ taskId, eventType, payload });
-  }
-  saveAttempt(attempt: Attempt): void {
-    const list = this.attempts.get(attempt.nodeId) ?? [];
-    list.push(attempt);
-    this.attempts.set(attempt.nodeId, list);
-  }
-  loadAttempts(nodeId: string): Attempt[] { return this.attempts.get(nodeId) ?? []; }
-  loadAttempt(attemptId: string): Attempt | undefined {
-    for (const list of this.attempts.values()) {
-      const found = list.find((a) => a.id === attemptId);
-      if (found) return found;
-    }
-    return undefined;
-  }
-  updateAttempt(
-    attemptId: string,
-    changes: Partial<Pick<Attempt,
-      | 'status' | 'startedAt' | 'completedAt' | 'exitCode' | 'error'
-      | 'lastHeartbeatAt' | 'branch' | 'commit' | 'summary'
-      | 'workspacePath' | 'agentSessionId' | 'containerId' | 'mergeConflict'>>,
-  ): void {
-    for (const list of this.attempts.values()) {
-      const idx = list.findIndex((a) => a.id === attemptId);
-      if (idx !== -1) {
-        list[idx] = { ...list[idx], ...changes } as Attempt;
-        return;
-      }
-    }
-  }
-  deleteWorkflow(workflowId: string): void {
-    this.workflows.delete(workflowId);
-    for (const [id, entry] of this.tasks) {
-      if (entry.workflowId === workflowId) this.tasks.delete(id);
-    }
-  }
-  deleteAllWorkflows(): void { this.workflows.clear(); this.tasks.clear(); }
-}
-
-class InMemoryBus implements OrchestratorMessageBus {
-  publish<T>(_channel: string, _message: T): void {}
-  subscribe(_channel: string, _handler: (msg: unknown) => void): () => void {
-    return () => undefined;
-  }
-}
-
-function makeOrchestrator(persistence: OrchestratorPersistence): Orchestrator {
-  return new Orchestrator({
-    persistence,
-    messageBus: new InMemoryBus(),
-    maxConcurrency: 8,
-  });
-}
+import {
+  InMemoryPersistence,
+  makeOrchestrator,
+  makeResponse,
+  setupChain,
+} from './helpers/cross-workflow-cascade-helpers.js';
 
 function buildOrchestratorDeps(orchestrator: Orchestrator): InvalidationDeps {
   return {
@@ -170,108 +50,6 @@ function buildOrchestratorDeps(orchestrator: Orchestrator): InvalidationDeps {
       if (!workflowId) return [];
       return orchestrator.cascadeInvalidationToDownstream(workflowId);
     },
-  };
-}
-
-function makeResponse(overrides: Partial<WorkResponse>): WorkResponse {
-  return {
-    requestId: 'req-1',
-    actionId: 't1',
-    executionGeneration: 0,
-    status: 'completed',
-    outputs: { exitCode: 0 },
-    ...overrides,
-  };
-}
-
-/**
- * Drive a chain of two workflows so the downstream first task has cleared the
- * external merge-gate dependency, runs to completion locally, and a downstream
- * later task is currently `running`. This mirrors the production scenario in
- * `~/.invoker/invoker.db` for WF-INV-113 → WF-INV-55 (see plan reproduction
- * notes) where the upstream merge gate cleared once, the downstream cascaded
- * forward on local deps, and an upstream recreate fails to invalidate
- * downstream work.
- */
-function setupChain(orchestrator: Orchestrator): {
-  upstreamWfId: string;
-  upstreamTaskId: string;
-  upstreamMergeId: string;
-  downstreamWfId: string;
-  downstreamRootId: string;
-  downstreamMidId: string;
-  downstreamLastId: string;
-  downstreamMergeId: string;
-} {
-  orchestrator.loadPlan({
-    name: 'upstream-workflow',
-    baseBranch: 'master',
-    featureBranch: 'feature/upstream',
-    tasks: [{ id: 'verify-upstream', description: 'upstream prerequisite' }],
-  });
-  const upstreamTaskId = orchestrator.getAllTasks().find(
-    (t) => !t.config.isMergeNode && t.id.endsWith('/verify-upstream'),
-  )!.id;
-  const upstreamWfId = upstreamTaskId.split('/')[0]!;
-  const upstreamMergeId = `__merge__${upstreamWfId}`;
-
-  orchestrator.loadPlan({
-    name: 'downstream-workflow',
-    baseBranch: 'feature/upstream',
-    featureBranch: 'feature/downstream',
-    tasks: [
-      {
-        id: 'root',
-        description: 'downstream root waits for upstream merge gate',
-        externalDependencies: [{ workflowId: upstreamWfId, gatePolicy: 'completed' }],
-      },
-      {
-        id: 'mid',
-        description: 'downstream mid depends on root',
-        dependencies: ['root'],
-      },
-      {
-        id: 'last',
-        description: 'downstream last depends on mid',
-        dependencies: ['mid'],
-      },
-    ],
-  });
-
-  const downstreamRootId = orchestrator.getAllTasks().find(
-    (t) => t.id.endsWith('/root'),
-  )!.id;
-  const downstreamWfId = downstreamRootId.split('/')[0]!;
-  const downstreamMidId = `${downstreamWfId}/mid`;
-  const downstreamLastId = `${downstreamWfId}/last`;
-  const downstreamMergeId = `__merge__${downstreamWfId}`;
-
-  // Drive the upstream task and merge gate to `completed` so the
-  // downstream root clears the external dependency and runs.
-  orchestrator.startExecution();
-  orchestrator.handleWorkerResponse(makeResponse({ actionId: upstreamTaskId, status: 'completed' }));
-  orchestrator.handleWorkerResponse(makeResponse({ actionId: upstreamMergeId, status: 'completed' }));
-
-  // Downstream root should now be running.
-  expect(orchestrator.getTask(downstreamRootId)!.status).toBe('running');
-
-  // Cascade downstream root → mid → last on local deps.
-  orchestrator.handleWorkerResponse(makeResponse({ actionId: downstreamRootId, status: 'completed' }));
-  orchestrator.handleWorkerResponse(makeResponse({ actionId: downstreamMidId, status: 'completed' }));
-
-  // Downstream `last` is now running locally; merge gate still pending.
-  expect(orchestrator.getTask(downstreamLastId)!.status).toBe('running');
-  expect(orchestrator.getTask(downstreamMergeId)!.status).toBe('pending');
-
-  return {
-    upstreamWfId,
-    upstreamTaskId,
-    upstreamMergeId,
-    downstreamWfId,
-    downstreamRootId,
-    downstreamMidId,
-    downstreamLastId,
-    downstreamMergeId,
   };
 }
