+++ packages/workflow-core/src/__tests__/orchestrator.test.ts
@@ -5869,7 +5869,7 @@ describe('Orchestrator', () => {
   // ── retryWorkflow ────────────────────────────────────────
 
   describe('retryWorkflow', () => {
-    it('refreshes only the targeted workflow before retrying', () => {
+    it('does not modify unrelated workflows when retrying', () => {
       const persistence = new CountingPersistence();
       const bus = new InMemoryBus();
       const o = new Orchestrator({
@@ -5892,13 +5892,17 @@ describe('Orchestrator', () => {
         ],
       });
 
-      persistence.loadTasksCalls = [];
       const wfA = o.getAllTasks().find((task) => task.id.endsWith('/a1'))!.config.workflowId!;
+      const wfBTaskBefore = o.getAllTasks().find((task) => task.id.endsWith('/b1'))!;
+
+      persistence.loadTasksCalls = [];
       o.retryWorkflow(wfA);
 
       expect(persistence.loadTasksCalls.filter((id) => id === wfA).length).toBeGreaterThan(0);
-      const wfB = o.getAllTasks().find((task) => task.id.endsWith('/b1'))!.config.workflowId!;
-      expect(persistence.loadTasksCalls).not.toContain(wfB);
+
+      const wfBTaskAfter = o.getAllTasks().find((task) => task.id.endsWith('/b1'))!;
+      expect(wfBTaskAfter.status).toBe(wfBTaskBefore.status);
+      expect(wfBTaskAfter.execution.generation ?? 0).toBe(wfBTaskBefore.execution.generation ?? 0);
     });
 
     it('preserves completed tasks and resets failed tasks', () => {
